LayoutTests/media/video-src-blob-replay-expected.txt

OPEN FILE PANEL
This tests the ability of the <video> element to load blob URLs, play to the end, and replay. In the browser, select a video file:

EVENT(change)
EVENT(playing)
EVENT(seeked)
END OF TEST
